+++ koala/review2/index.tsx
@@ -1,4 +1,4 @@
-import { Box, Stack } from "@mantine/core";
+import { Box } from "@mantine/core";
 import React from "react";
 import { Feedback } from "./lesson-steps/feedback";
 import { Listening } from "./lesson-steps/listening";
@@ -8,7 +8,7 @@ import { Pending } from "./lesson-steps/pending";
 import { RemedialIntro } from "./lesson-steps/remedial-intro";
 import { RemedialOutro } from "./lesson-steps/remedial-outro";
 import { Speaking } from "./lesson-steps/speaking";
-import { TopBar } from "./top-bar";
+import { ControlBar } from "./control-bar";
 import { CardReviewProps, CardUI, ItemType } from "./types";
 
 const cardUIs: Record<ItemType, CardUI> = {
@@ -39,15 +39,34 @@ export const CardReview: React.FC<CardReviewWithRecordingProps> = (
   const CardComponent = cardUIs[itemType] ?? UnknownCard;
 
   return (
-    <Stack>
-      <TopBar
-        {...props}
-        onRecordingComplete={onRecordingComplete}
-        currentStepUuid={currentStepUuid}
-      />
-      <Box>
+    <Box
+      style={{
+        minHeight: "100vh",
+        position: "relative",
+        paddingBottom: "80px",
+      }}
+    >
+      <Box p="md">
         <CardComponent {...props} />
       </Box>
-    </Stack>
+      <Box
+        style={{
+          position: "fixed",
+          bottom: 0,
+          left: 0,
+          right: 0,
+          backgroundColor: "var(--mantine-color-pink-0)",
+          borderTop: "1px solid var(--mantine-color-pink-2)",
+          padding: "var(--mantine-spacing-sm)",
+          zIndex: 100,
+        }}
+      >
+        <ControlBar
+          {...props}
+          onRecordingComplete={onRecordingComplete}
+          currentStepUuid={currentStepUuid}
+        />
+      </Box>
+    </Box>
   );
 };
